relocate cunstom_dataset, deal with pcolor warning about non-monotonical increasing
