+++ src/utils/__init__.py
@@ -1,10 +1,8 @@
 # only for usage of `from src.utils import *`
-__all__ = ["DataCompose", "DataType", "Level", "ModelType", "gen_path", "gen_data"]
+__all__ = ["DataCompose", "DataType", "Level", "gen_path", "gen_data"]
 
 # define package members, be careful of circular import
-from .custom_dataset import *
 from .data_compose import *
 from .data_type import *
 from .file_util import *
-from .model_type import *
 from .time_util import *
