--- src/utils/model_type.py
@@ -1,5 +1,0 @@
-from enum import StrEnum
-
-
-class ModelType(StrEnum):
-    Pangu = "Pangu-weather"
